+++ tests/BotGenerator.spec.ts
@@ -64,6 +64,7 @@ describe('teams:bot', function () {
     it('should generate bot project with v1.3 without unit tests', async () => {
         await helpers.run(testHelper.GENERATOR_PATH)
             .inDir(testHelper.TEMP_BOT_GENERATOR_PATH + '/bot02')
+            .withArguments('[no-telementry]')
             .withPrompts({
                 solutionName: 'bot-test-01',
                 whichFolder: 'current',
@@ -92,6 +93,7 @@ describe('teams:bot', function () {
     it('should generate bot project with devPreview with unit tests', async () => {
         await helpers.run(testHelper.GENERATOR_PATH)
             .inDir(testHelper.TEMP_BOT_GENERATOR_PATH + '/bot03')
+            .withArguments('[no-telementry]')
             .withPrompts({
                 solutionName: 'bot-test-01',
                 whichFolder: 'current',
@@ -120,6 +122,7 @@ describe('teams:bot', function () {
     it('should generate bot project with devPreview without unit tests', async () => {
         await helpers.run(testHelper.GENERATOR_PATH)
             .inDir(testHelper.TEMP_BOT_GENERATOR_PATH + '/bot04')
+            .withArguments('[no-telementry]')
             .withPrompts({
                 solutionName: 'bot-test-01',
                 whichFolder: 'current',