+++ tests/ConnectorGenerator.spec.ts
@@ -61,6 +61,7 @@ describe('teams:connector', function () {
     it('should generate connector project with v1.3 without unit tests', async () => {
         await helpers.run(testHelper.GENERATOR_PATH)
             .inDir(testHelper.TEMP_CONNECTOR_GENERATOR_PATH + '/connector02')
+            .withArguments('[no-telementry]')
             .withPrompts({
                 solutionName: 'connector-test-01',
                 whichFolder: 'current',
@@ -89,6 +90,7 @@ describe('teams:connector', function () {
     it('should generate connector project with devPreview with unit tests', async () => {
         await helpers.run(testHelper.GENERATOR_PATH)
             .inDir(testHelper.TEMP_CONNECTOR_GENERATOR_PATH + '/connector03')
+            .withArguments('[no-telementry]')
             .withPrompts({
                 solutionName: 'connector-test-01',
                 whichFolder: 'current',
@@ -117,6 +119,7 @@ describe('teams:connector', function () {
     it('should generate connector project with devPreview without unit tests', async () => {
         await helpers.run(testHelper.GENERATOR_PATH)
             .inDir(testHelper.TEMP_CONNECTOR_GENERATOR_PATH + '/connector04')
+            .withArguments('[no-telementry]')
             .withPrompts({
                 solutionName: 'connector-test-01',
                 whichFolder: 'current',